move sentry user id generation util to sentry/user
