+++ src/utils/user/index.ts
@@ -5,29 +5,7 @@ import {
 } from 'types/user';
 import { getActualKey } from 'utils/common/key';
 import CryptoWorker from 'utils/crypto';
-import { getData, LS_KEYS, setData } from 'utils/storage/localStorage';
-
-export function makeID(length) {
-    let result = '';
-    const characters =
-        'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-    const charactersLength = characters.length;
-    for (let i = 0; i < length; i++) {
-        result += characters.charAt(
-            Math.floor(Math.random() * charactersLength)
-        );
-    }
-    return result;
-}
-
-export function getUserAnonymizedID() {
-    let anonymizeUserID = getData(LS_KEYS.AnonymizeUserID)?.id;
-    if (!anonymizeUserID) {
-        anonymizeUserID = makeID(6);
-        setData(LS_KEYS.AnonymizeUserID, { id: anonymizeUserID });
-    }
-    return anonymizeUserID;
-}
+import { getData, LS_KEYS } from 'utils/storage/localStorage';
 
 export function getLocalUserDetails(): UserDetails {
     return getData(LS_KEYS.USER_DETAILS)?.value;
